Extract the Hapi Server to expose it for testing
* The Hapi Server has been extracted so that it can be
tested using the built in #inject() method.
* Test support has been moved to /helpers. This is the default location
that is ignored by `ava`.
* Added a test to ensure that the user can override the default html.js
in both the develop and build stages.

After some research I think using Hapi's inject is going to provide the
best testing experience. The only code change was that the server is now
started in `lib/utils/develop` and constructed in `lib/utils/dev-server`
