+++ package.json
@@ -82,7 +82,6 @@
   },
   "devDependencies": {
     "ava": "^0.14.0",
-    "ava-http": "^0.2.1",
     "babel-cli": "^6.7.5",
     "babel-eslint": "^6.0.2",
     "babel-register": "^6.7.2",
