+++ test/fixtures/override-htmljs/pages/index.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function index () {
+  return <h1>Index</h1>
+}
